fix: 修复最小发送间隔支持小数
- setup.py: int -> float
- proactive/config.py: min_interval_hours 返回类型 int -> float

+++ ai_companion/proactive/config.py
@@ -114,7 +114,7 @@ class ProactiveConfig:
         return self._config.get("scheduler", {}).get("max_daily", 5)
 
     @property
-    def min_interval_hours(self) -> int:
+    def min_interval_hours(self) -> float:
         return self._config.get("scheduler", {}).get("min_interval_hours", 4)
 
     @property
